# docs/swagger/bloodconnect.yaml
openapi: 3.0.3
info:
  title: BloodConnect API
  description: |
    API-first contract for BloodConnect REST API.

    This document is the shared source of truth for Backend, Flutter, Postman,
    Swagger UI, and future client SDK generation.
  version: 1.0.0
  contact:
    name: BloodConnect API Team
  license:
    name: Proprietary
    url: https://bloodconnect.id/license
servers:
  - url: /api/v1
    description: Base API path
security:
  - bearerAuth: []
...
paths:
...
  /auth/change-password:
    put:
      tags:
        - Authentication
      summary: Change authenticated user password
      description: Changes the password of the currently authenticated user.
      operationId: changePassword
      requestBody:
        $ref: "#/components/requestBodies/ChangePasswordRequestBody"
      responses:
        "200":
          $ref: "#/components/responses/EmptySuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /blood-requests:
    get:
      tags:
        - Blood Request
      summary: List blood requests
      description: Returns blood requests visible to the authenticated user.
      operationId: listBloodRequests
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/BloodTypeParam"
        - $ref: "#/components/parameters/RhesusParam"
        - $ref: "#/components/parameters/UrgencyParam"
        - $ref: "#/components/parameters/RequestStatusParam"
      responses:
        "200":
          description: Blood requests retrieved successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BloodRequestListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Blood Request
      summary: Create blood request
      description: Creates a new blood request for a patient.
      operationId: createBloodRequest
      requestBody:
        $ref: "#/components/requestBodies/CreateBloodRequestBody"
      responses:
        "201":
          description: Blood request created successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BloodRequestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /blood-requests/{blood_request_id}:
    get:
      tags:
        - Blood Request
      summary: Get blood request detail
      description: Returns a specific blood request including requester information.
      operationId: getBloodRequestById
      parameters:
        - $ref: "#/components/parameters/BloodRequestIdPath"
      responses:
        "200":
          $ref: "#/components/responses/BloodRequestSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags:
        - Blood Request
      summary: Update blood request
      description: Updates an owned blood request while it is still editable.
      operationId: updateBloodRequest
      parameters:
        - $ref: "#/components/parameters/BloodRequestIdPath"
      requestBody:
        $ref: "#/components/requestBodies/UpdateBloodRequestBody"
      responses:
        "200":
          $ref: "#/components/responses/BloodRequestSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Blood Request
      summary: Delete blood request
      description: Soft deletes an owned blood request.
      operationId: deleteBloodRequest
      parameters:
        - $ref: "#/components/parameters/BloodRequestIdPath"
      responses:
        "200":
          $ref: "#/components/responses/EmptySuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token using the Authorization header.
        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIs...`
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    BloodRequestIdPath:
      name: blood_request_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BloodTypeParam:
      name: blood_type
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/BloodType"
    RhesusParam:
      name: rhesus
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Rhesus"
    UrgencyParam:
      name: urgency
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Urgency"
    RequestStatusParam:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/RequestStatus"
...
  requestBodies:
...
    ChangePasswordRequestBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ChangePasswordRequest"
...
    CreateBloodRequestBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CreateBloodRequest"
    UpdateBloodRequestBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UpdateBloodRequest"
...
  responses:
    EmptySuccess:
      description: Operation completed successfully.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessResponse"
          examples:
            success:
              $ref: "#/components/examples/EmptySuccessExample"
...
    BloodRequestSuccess:
      description: Blood request retrieved successfully.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/SuccessResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/BloodRequestResponse"
...
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            badRequest:
              $ref: "#/components/examples/BadRequestExample"
    Unauthorized:
      description: Unauthorized request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              $ref: "#/components/examples/UnauthorizedExample"
    Forbidden:
      description: Forbidden request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            forbidden:
              $ref: "#/components/examples/ForbiddenExample"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              $ref: "#/components/examples/NotFoundExample"
...
    ValidationError:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          examples:
            validationError:
              $ref: "#/components/examples/ValidationErrorExample"
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internalServerError:
              $ref: "#/components/examples/InternalServerErrorExample"
  schemas:
    Gender:
      type: string
      enum:
        - Male
        - Female
    BloodType:
      type: string
      enum:
        - A
        - B
        - AB
        - O
    Rhesus:
      type: string
      enum:
        - "+"
        - "-"
    Urgency:
      type: string
      enum:
        - Low
        - Medium
        - High
        - Critical
    RequestStatus:
      type: string
      enum:
        - Pending
        - Searching
        - Donor Found
        - Completed
        - Cancelled
...
    ChangePasswordRequest:
      type: object
      required:
        - old_password
        - new_password
        - confirm_password
      properties:
        old_password:
          type: string
          format: password
          writeOnly: true
        new_password:
          type: string
          format: password
          minLength: 8
          writeOnly: true
        confirm_password:
          type: string
          format: password
          minLength: 8
          writeOnly: true
...
    CreateBloodRequest:
      type: object
      required:
        - patient_name
        - hospital_name
        - hospital_address
        - latitude
        - longitude
        - blood_type
        - rhesus
        - bags_needed
        - urgency
      properties:
        patient_name:
          type: string
          maxLength: 100
        relationship:
          type: string
          maxLength: 50
          nullable: true
        hospital_name:
          type: string
          maxLength: 150
        hospital_address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        blood_type:
          $ref: "#/components/schemas/BloodType"
        rhesus:
          $ref: "#/components/schemas/Rhesus"
        bags_needed:
          type: integer
          minimum: 1
        urgency:
          $ref: "#/components/schemas/Urgency"
        note:
          type: string
          nullable: true
    UpdateBloodRequest:
      allOf:
        - $ref: "#/components/schemas/CreateBloodRequest"
...
    UserResponse:
      type: object
      required:
        - id
        - full_name
        - email
        - phone
        - blood_type
        - rhesus
        - is_available
        - is_verified
        - is_active
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
          example: Andi Pratama
        email:
          type: string
          format: email
        phone:
          type: string
        gender:
          $ref: "#/components/schemas/Gender"
        birth_date:
          type: string
          format: date
          nullable: true
        blood_type:
          $ref: "#/components/schemas/BloodType"
        rhesus:
          $ref: "#/components/schemas/Rhesus"
        weight:
          type: integer
          nullable: true
        last_donor_date:
          type: string
          format: date
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        profile_photo:
          type: string
          format: uri
          nullable: true
        is_available:
          type: boolean
        is_verified:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    BloodRequestResponse:
      type: object
      required:
        - id
        - user_id
        - patient_name
        - hospital_name
        - hospital_address
        - latitude
        - longitude
        - blood_type
        - rhesus
        - bags_needed
        - urgency
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        patient_name:
          type: string
        relationship:
          type: string
          nullable: true
        hospital_name:
          type: string
        hospital_address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        blood_type:
          $ref: "#/components/schemas/BloodType"
        rhesus:
          $ref: "#/components/schemas/Rhesus"
        bags_needed:
          type: integer
        urgency:
          $ref: "#/components/schemas/Urgency"
        note:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/RequestStatus"
        fulfilled_at:
          type: string
          format: date-time
          nullable: true
        requester:
          $ref: "#/components/schemas/UserResponse"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BloodRequestListResponse:
      type: object
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BloodRequestResponse"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
...
    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 42
        total_pages:
          type: integer
          example: 5
    SuccessResponse:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      required:
        - success
        - message
        - errors
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Request failed
        errors:
          type: object
          nullable: true
          additionalProperties: true
    ValidationErrorResponse:
      type: object
      required:
        - success
        - message
        - errors
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  examples:
...
    EmptySuccessExample:
      summary: Empty success response
      value:
        success: true
        message: Operation completed successfully
        data: {}
    BadRequestExample:
      summary: 400 Bad Request
      value:
        success: false
        message: Invalid request body
        errors:
          body:
            - Request body must be valid JSON
    UnauthorizedExample:
      summary: 401 Unauthorized
      value:
        success: false
        message: Unauthorized
        errors:
          token:
            - Token is missing, invalid, or expired
    ForbiddenExample:
      summary: 403 Forbidden
      value:
        success: false
        message: Access forbidden
        errors:
          authorization:
            - You do not have permission to access this resource
    NotFoundExample:
      summary: 404 Not Found
      value:
        success: false
        message: Resource not found
        errors:
          resource:
            - The requested resource was not found
...
    ValidationErrorExample:
      summary: 422 Unprocessable Entity
      value:
        success: false
        message: Validation failed
        errors:
          email:
            - Email is required
          password:
            - Password must be at least 8 characters
    InternalServerErrorExample:
      summary: 500 Internal Server Error
      value:
        success: false
        message: Internal server error
        errors:
          server:
            - An unexpected error occurred